Integration: stdio helper binary and move helpers.
Move the inline MCP stdio server compiled by call_tool_stdio.go into a
reusable helper binary at tests/integration/framework/binary/helpers/
mcpstdioserver, prebuilt by BuildAll. The test now references the helper
via binary.EnvValue and asserts the workflow output JSON end-to-end with
assert.JSONEq.

Move MCP helper funcs into framework.

cc @sicoyle
