+++ tests/integration/framework/binary/helpers/mcpstdioserver/go.mod
@@ -1,0 +1,14 @@
+module mcpstdioserver
+
+go 1.26.3
+
+require github.com/modelcontextprotocol/go-sdk v1.6.0
+
+require (
+	github.com/google/jsonschema-go v0.4.3 // indirect
+	github.com/segmentio/asm v1.1.3 // indirect
+	github.com/segmentio/encoding v0.5.4 // indirect
+	github.com/yosida95/uritemplate/v3 v3.0.2 // indirect
+	golang.org/x/oauth2 v0.35.0 // indirect
+	golang.org/x/sys v0.41.0 // indirect
+)
