+++ tests/integration/suite/daprd/mcpserver/call_tool_per_tool_workflow.go
@@ -32,6 +32,7 @@ import (
 	prochttp "github.com/dapr/dapr/tests/integration/framework/process/http"
 	"github.com/dapr/dapr/tests/integration/framework/process/placement"
 	"github.com/dapr/dapr/tests/integration/framework/process/scheduler"
+	"github.com/dapr/dapr/tests/integration/framework/workflow/httpapi"
 	"github.com/dapr/dapr/tests/integration/suite"
 )
 
@@ -121,9 +122,9 @@ func (s *callToolPerToolWorkflow) Run(t *testing.T, ctx context.Context) {
 			"arguments": map[string]any{"city": "Portland"},
 		}
 		// Use the per-tool workflow name: dapr.internal.mcp.multi-tool.CallTool.get_weather
-		status := runWorkflow(t, ctx, s.httpClient, s.daprd.HTTPPort(),
+		status := httpapi.Run(t, ctx, s.httpClient, s.daprd.HTTPPort(),
 			mcpnames.MCPCallToolWorkflowName("multi-tool", "get_weather"), input, 30*time.Second)
-		require.Equal(t, statusCompleted, status.RuntimeStatus)
+		require.Equal(t, httpapi.StatusCompleted, status.RuntimeStatus)
 
 		outputJSON := status.Properties["dapr.workflow.output"]
 		require.NotEmpty(t, outputJSON)
@@ -140,9 +141,9 @@ func (s *callToolPerToolWorkflow) Run(t *testing.T, ctx context.Context) {
 			"arguments": map[string]any{"name": "Dapr"},
 		}
 		// Use the greet tool's workflow name
-		status := runWorkflow(t, ctx, s.httpClient, s.daprd.HTTPPort(),
+		status := httpapi.Run(t, ctx, s.httpClient, s.daprd.HTTPPort(),
 			mcpnames.MCPCallToolWorkflowName("multi-tool", "greet"), input, 30*time.Second)
-		require.Equal(t, statusCompleted, status.RuntimeStatus)
+		require.Equal(t, httpapi.StatusCompleted, status.RuntimeStatus)
 
 		outputJSON := status.Properties["dapr.workflow.output"]
 		require.NotEmpty(t, outputJSON)
@@ -159,9 +160,9 @@ func (s *callToolPerToolWorkflow) Run(t *testing.T, ctx context.Context) {
 		// StartWorkflow accepts the call; the workflow itself transitions to
 		// FAILED with "is not registered" in the failure details.
 		input := map[string]any{"arguments": map[string]any{}}
-		status := runWorkflow(t, ctx, s.httpClient, s.daprd.HTTPPort(),
+		status := httpapi.Run(t, ctx, s.httpClient, s.daprd.HTTPPort(),
 			mcpnames.MCPCallToolWorkflowName("multi-tool", "nonexistent_tool"), input, 30*time.Second)
-		require.Equal(t, statusFailed, status.RuntimeStatus)
+		require.Equal(t, httpapi.StatusFailed, status.RuntimeStatus)
 		assert.Contains(t, status.Properties["dapr.workflow.failure.error_message"], "is not registered")
 	})
 
@@ -171,9 +172,9 @@ func (s *callToolPerToolWorkflow) Run(t *testing.T, ctx context.Context) {
 		input := map[string]any{
 			"arguments": map[string]any{"city": "Seattle"},
 		}
-		status := runWorkflow(t, ctx, s.httpClient, s.daprd.HTTPPort(),
+		status := httpapi.Run(t, ctx, s.httpClient, s.daprd.HTTPPort(),
 			mcpnames.MCPCallToolWorkflowName("multi-tool", "get_weather"), input, 30*time.Second)
-		require.Equal(t, statusCompleted, status.RuntimeStatus,
+		require.Equal(t, httpapi.StatusCompleted, status.RuntimeStatus,
 			"should succeed using tool name from workflow name suffix")
 
 		outputJSON := status.Properties["dapr.workflow.output"]
@@ -187,9 +188,9 @@ func (s *callToolPerToolWorkflow) Run(t *testing.T, ctx context.Context) {
 	})
 
 	t.Run("ListTools still returns all tools", func(t *testing.T) {
-		status := runWorkflow(t, ctx, s.httpClient, s.daprd.HTTPPort(),
+		status := httpapi.Run(t, ctx, s.httpClient, s.daprd.HTTPPort(),
 			mcpnames.MCPListToolsWorkflowName("multi-tool"), nil, 30*time.Second)
-		require.Equal(t, statusCompleted, status.RuntimeStatus)
+		require.Equal(t, httpapi.StatusCompleted, status.RuntimeStatus)
 
 		outputJSON := status.Properties["dapr.workflow.output"]
 		require.NotEmpty(t, outputJSON)
