+++ tests/integration/suite/daprd/mcpserver/hotreload_tool_change.go
@@ -34,6 +34,7 @@ import (
 	prochttp "github.com/dapr/dapr/tests/integration/framework/process/http"
 	"github.com/dapr/dapr/tests/integration/framework/process/placement"
 	"github.com/dapr/dapr/tests/integration/framework/process/scheduler"
+	"github.com/dapr/dapr/tests/integration/framework/workflow/httpapi"
 	"github.com/dapr/dapr/tests/integration/suite"
 )
 
@@ -118,10 +119,10 @@ func (s *hotReloadToolChange) Run(t *testing.T, ctx context.Context) {
 
 	t.Run("initial: alpha tool workflow works", func(t *testing.T) {
 		input := map[string]any{"arguments": map[string]any{}}
-		status, err := tryRunWorkflow(ctx, s.httpClient, s.daprd.HTTPPort(),
+		status, err := httpapi.TryRun(ctx, s.httpClient, s.daprd.HTTPPort(),
 			mcpnames.MCPCallToolWorkflowName("dynamic", "alpha"), input, 30*time.Second)
 		require.NoError(t, err)
-		require.Equal(t, statusCompleted, status.RuntimeStatus)
+		require.Equal(t, httpapi.StatusCompleted, status.RuntimeStatus)
 
 		var result mcp.CallToolResult
 		require.NoError(t, json.Unmarshal([]byte(status.Properties["dapr.workflow.output"]), &result))
@@ -134,19 +135,19 @@ func (s *hotReloadToolChange) Run(t *testing.T, ctx context.Context) {
 	t.Run("after hot-reload: beta tool workflow works", func(t *testing.T) {
 		input := map[string]any{"arguments": map[string]any{}}
 		assert.EventuallyWithT(t, func(c *assert.CollectT) {
-			status, err := tryRunWorkflow(ctx, s.httpClient, s.daprd.HTTPPort(),
+			status, err := httpapi.TryRun(ctx, s.httpClient, s.daprd.HTTPPort(),
 				mcpnames.MCPCallToolWorkflowName("dynamic", "beta"), input, 5*time.Second)
 			if !assert.NoError(c, err) {
 				return
 			}
-			assert.Equal(c, statusCompleted, status.RuntimeStatus)
+			assert.Equal(c, httpapi.StatusCompleted, status.RuntimeStatus)
 		}, 30*time.Second, time.Second)
 	})
 
 	t.Run("after hot-reload: alpha tool workflow unregistered", func(t *testing.T) {
 		// alpha is no longer registered after hot-reload to server B. The runtime
 		// surfaces this in one of two ways depending on platform/timing:
-		//   - WaitForInstanceStart blocks → tryRunWorkflow's per-tick timeout
+		//   - WaitForInstanceStart blocks → httpapi.TryRun's per-tick timeout
 		//     returns a polling error.
 		//   - The schedule succeeds and the workflow immediately completes with
 		//     FAILED (durabletask "no handler" fallback).
@@ -154,13 +155,13 @@ func (s *hotReloadToolChange) Run(t *testing.T, ctx context.Context) {
 		// successful COMPLETED status (which would mean alpha is still wired).
 		input := map[string]any{"arguments": map[string]any{}}
 		assert.EventuallyWithT(t, func(c *assert.CollectT) {
-			status, err := tryRunWorkflow(ctx, s.httpClient, s.daprd.HTTPPort(),
+			status, err := httpapi.TryRun(ctx, s.httpClient, s.daprd.HTTPPort(),
 				mcpnames.MCPCallToolWorkflowName("dynamic", "alpha"), input, 3*time.Second)
 			if err != nil {
 				// Polling error (start hung) — alpha is gone. Pass.
 				return
 			}
-			assert.NotEqual(c, statusCompleted, status.RuntimeStatus,
+			assert.NotEqual(c, httpapi.StatusCompleted, status.RuntimeStatus,
 				"alpha workflow should not complete successfully after hot-reload to server B")
 		}, 30*time.Second, time.Second)
 	})
